server/requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic
pydantic-settings
python-dotenv
